feat: fix templates list and rename variables
